TINY-4366: Fixed a circular dependency
