+++ modules/alloy/src/main/ts/ephox/alloy/dragging/common/DragMovement.ts
@@ -1,5 +1,5 @@
 import { Option } from '@ephox/katamari';
-import { Css, Scroll, Traverse } from '@ephox/sugar';
+import { Css, Element, Location, Scroll, Traverse } from '@ephox/sugar';
 
 import { cap } from '../../alien/Cycles';
 import * as OffsetOrigin from '../../alien/OffsetOrigin';
@@ -8,7 +8,23 @@ import { AlloyComponent } from '../../api/component/ComponentApi';
 import * as DragCoord from '../../api/data/DragCoord';
 import * as Snappables from '../snap/Snappables';
 import { DraggingConfig, DragStartData, SnapsConfig } from './DraggingTypes';
-import * as DragUtils from './DragUtils';
+
+const getCurrentCoord = (target: Element): DragCoord.CoordAdt => {
+  return Css.getRaw(target, 'left').bind((left) => {
+    return Css.getRaw(target, 'top').bind((top) => {
+      return Css.getRaw(target, 'position').map((position) => {
+        const nu = position === 'fixed' ? DragCoord.fixed : DragCoord.offset;
+        return nu(
+          parseInt(left, 10),
+          parseInt(top, 10)
+        );
+      });
+    });
+  }).getOrThunk(() => {
+    const location = Location.absolute(target);
+    return DragCoord.absolute(location.left(), location.top());
+  });
+};
 
 const clampCoords = (component: AlloyComponent, coords: DragCoord.CoordAdt, scroll: SugarPosition, origin: SugarPosition, startData: DragStartData): DragCoord.CoordAdt => {
   const bounds = startData.bounds;
@@ -62,7 +78,7 @@ const dragBy = (component: AlloyComponent, dragConfig: DraggingConfig, startData
 
     const origin = OffsetOrigin.getOrigin(target);
 
-    const currentCoord = DragUtils.getCurrentCoord(target);
+    const currentCoord = getCurrentCoord(target);
 
     const newCoord = calcNewCoord(component, dragConfig.snaps, currentCoord, scroll, origin, delta, startData);
